amended bnn script to do evaluation properly using 25 prediction samples debug - comparing groundtruths to mean predictions

+++ bayesian_modeltrain.py
@@ -132,7 +132,7 @@ def evaluate_regression(model, valid_iterator, val_set, loss_fn, n_samples=25,  
     ci_acc = ci_acc.float().mean()
     r2 = r2_score(means.detach().numpy(), gt.detach().numpy())
     r = pearsonr(means.detach().numpy().ravel(), gt.detach().numpy().ravel())[0]
-    return loss.detach().numpy(), r, r2, ci_acc, (ci_upper >= gt).mean(), (ci_lower <= gt).mean()
+    return loss.detach().numpy(), r, r2, ci_acc, (ci_upper >= gt).float().mean(), (ci_lower <= gt).float().mean()
